requirements.txt
tkinter
customtkinter
pandas
numpy
pickle-mixin
mediapipe
opencv-python
Pillow
flask
flask-cors
scikit-learn
